requirements.txt
absl-py==1.4.0
accelerate==0.22.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.0.0
altair==5.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==22.1.0
av==10.0.0
Babel==2.12.1
backcall==0.2.0
batch-face==1.4.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
black==23.7.0
bleach==5.0.1
boto3==1.28.44
botocore==1.31.44
brotlipy==0.7.0
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.26.4
comm==0.1.4
ConfigArgParse==1.7
contextlib2==21.6.0
contourpy==1.1.0
controlnet-aux==0.0.6
cryptography==41.0.3
cycler==0.11.0
Cython==0.29.34
datasets==2.14.4
debugpy==1.6.4
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.20.2
dill==0.3.7
docker-pycreds==0.4.0
dominate==2.7.0
easydict==1.10
einops==0.6.1
entrypoints==0.4
executing==1.2.0
face-alignment==1.4.0
facexlib==0.3.0
fairscale==0.4.6
fastapi==0.101.1
fastjsonschema==2.16.2
ffmpy==0.3.1
filelock==3.12.3
filterpy==1.4.5
flix-densepose @ file:///home/ubuntu/projects/python/vatsal/p2p_modules/packages/flix_densepose-0.1.0.tar.gz#sha256=2f0183830ae3b265b06d40e89cd6fd63f9417fba9c2b92d7455057c938c9e3f1
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gradio==3.41.0
gradio_client==0.5.0
grpcio==1.57.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hydra-core==1.3.2
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
iopath==0.1.9
ipdb==0.13.13
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.24.0
kiwisolver==1.4.4
kornia==0.7.0
lazy_loader==0.3
lit==16.0.6
llvmlite==0.39.1
lmdb==1.4.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
MegEngine==1.13.1+cu114
megfile==2.0.5
mistune==2.0.4
ml-collections==0.1.1
mmcv==2.0.1
mmcv-full==1.7.1
mmedit==0.16.1
mmengine==0.8.4
moviepy==1.0.3
mpmath==1.3.0
mprop==0.16.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
orjson==3.9.5
overrides==7.4.0
packaging==23.1
packbits==0.6
pandas==2.0.3
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
portalocker==2.7.0
POT==0.9.0
proglog==0.1.10
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==4.24.2
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
Pygments==2.13.0
PyNaCl==1.5.0
pyopencl==2023.1.2
pyparsing==3.1.1
pyre-extensions==0.0.29
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytools==2023.1.1
pytoshop==1.0.1
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
referencing==0.30.2
regex==2023.8.8
requests==2.28.2
requests-oauthlib==1.3.1
retinaface-pytorch==0.0.8
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.10.0
rsa==4.9
s3transfer==0.6.2
safetensors==0.3.2
scikit-image==0.21.0
scikit-learn==1.0.2
scipy==1.9.3
semantic-version==2.10.0
Send2Trash==1.8.0
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SSIM-PIL==1.0.14
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.3.0
terminado==0.17.0
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.9.5
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==0.9.1
torchvision==0.15.2
tornado==6.2
tqdm==4.66.1
traitlets==5.6.0
transformers==4.32.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
urllib3==1.26.15
uvicorn==0.23.2
wandb==0.15.9
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.4.2
websockets==11.0.3
Werkzeug==2.3.7
widgetsnbextension==4.0.3
wrapt==1.15.0
xformers==0.0.20
xxhash==3.3.0
yacs==0.1.8
yapf==0.40.1
yarl==1.9.2
zipp==3.15.0
